remote_gui/requirements.txt
qrcode
Pillow
mss
pynput
